+++ src/app/modules/dashboard/index.ts
@@ -1,0 +1,1 @@
+export * from './dashboard.component';
