+++ src/app/modules/hero/detail/hero-detail.component.css
