+++ src/app/modules/hero/index.ts
@@ -1,0 +1,4 @@
+export * from './hero.service';
+export * from './detail/hero-detail.component';
+export * from './list/heroes-list.component';
+export * from './search/hero-search.component';
