+++ src/app/modules/hero/list/heroes-list.component.ts
@@ -1,15 +1,16 @@
 import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 
-import { Hero } from './hero';
-import { HeroService } from './hero.service';
+
+import { HeroService } from '../hero.service';
+import { Hero } from '../../../viewmodels/hero/hero.model';
 
 @Component({
   selector: 'my-heroes',
-  templateUrl: './heroes.component.html',
-  styleUrls: ['./heroes.component.css']
+  templateUrl: './heroes-list.component.html',
+  styleUrls: ['./heroes-list.component.css']
 })
-export class HeroesComponent implements OnInit {
+export class HeroesListComponent implements OnInit {
   heroes: Hero[];
   selectedHero: Hero;
   addingHero = false;
