+++ src/app/modules/hero/search/hero-search.service.ts
@@ -5,8 +5,8 @@ import { Observable } from 'rxjs/Observable';
 import 'rxjs/add/operator/map';
 import 'rxjs/add/operator/catch';
 import 'rxjs/add/observable/throw';
+import { Hero } from '../../../viewmodels/hero/hero.model';
 
-import { Hero } from './hero';
 
 @Injectable()
 export class HeroSearchService {
