+++ src/app/viewmodels/hero/hero.model.ts
